refactor(cli): fix eslint errors
